+++ NINA/Utility/XISF.cs
@@ -590,8 +590,10 @@ namespace NINA.Utility {
                 AddImageProperty(XISFImageProperty.Instrument.Sensor.Temperature, metaData.Camera.Temperature.ToString(CultureInfo.InvariantCulture), "[degC] CCD temperature");
             }
             if (!double.IsNaN(metaData.Camera.PixelSize)) {
-                AddImageProperty(XISFImageProperty.Instrument.Sensor.XPixelSize, metaData.Camera.PixelSize.ToString(CultureInfo.InvariantCulture), "[um] Pixel X axis size");
-                AddImageProperty(XISFImageProperty.Instrument.Sensor.YPixelSize, metaData.Camera.PixelSize.ToString(CultureInfo.InvariantCulture), "[um] Pixel Y axis size");
+                double pixelX = metaData.Camera.PixelSize * Math.Max(metaData.Camera.BinX, 1);
+                double pixelY = metaData.Camera.PixelSize * Math.Max(metaData.Camera.BinY, 1);
+                AddImageProperty(XISFImageProperty.Instrument.Sensor.XPixelSize, pixelX.ToString(CultureInfo.InvariantCulture), "[um] Pixel X axis size");
+                AddImageProperty(XISFImageProperty.Instrument.Sensor.YPixelSize, pixelY.ToString(CultureInfo.InvariantCulture), "[um] Pixel Y axis size");
             }
 
             /* Observer */